+++ app/models/school_data_importer.rb
@@ -44,6 +44,7 @@ class SchoolDataImporter
     school.close_date = row.fetch("CloseDate")
     school.establishment_number = row.fetch("EstablishmentNumber")
     school.statutory_high_age = row.fetch("StatutoryHighAge")
+    school.phone_number = row.fetch("TelephoneNum")
     school
   end
 end