+++ spec/models/claim_spec.rb
@@ -331,6 +331,17 @@ RSpec.describe Claim, type: :model do
     end
   end
 
+  describe "#school" do
+    it "returns the current_school of the claim eligiblity" do
+      claim = Claim.new(eligibility: StudentLoans::Eligibility.new(current_school: schools(:penistone_grammar_school)))
+      expect(claim.school).to eq schools(:penistone_grammar_school)
+    end
+
+    it "returns nil if no eligibility is set" do
+      expect(Claim.new.school).to be_nil
+    end
+  end
+
   describe "#submit!" do
     around do |example|
       freeze_time { example.run }
@@ -440,10 +451,6 @@ RSpec.describe Claim, type: :model do
 
       expect(create(:claim, :approved).approvable?).to eq false
       expect(create(:claim, :rejected).approvable?).to eq false
-    end
-
-    it "returns false for a claim that doesn't have a confirmed identity" do
-      expect(build(:claim, :unverified).approvable?).to eq false
     end
   end